+++ stactask/__init__.py
@@ -6,6 +6,7 @@ except PackageNotFoundError:
     # package is not installed
     pass
 
+from .config import DownloadConfig
 from .task import Task
 
-__all__ = ["Task"]
+__all__ = ["Task", "DownloadConfig"]