+++ tests/test_task_download.py
@@ -1,8 +1,12 @@
 import json
+import os
 from pathlib import Path
 from typing import Any
 
 import pytest
+import stac_asset
+
+from stactask.config import DownloadConfig
 
 from .tasks import NothingTask
 
@@ -17,22 +21,47 @@ def item_collection() -> dict[str, Any]:
     return items
 
 
-def test_download_nosuch_asset(item_collection: dict[str, Any]) -> None:
+def test_download_nosuch_asset(tmp_path: Path, item_collection: dict[str, Any]) -> None:
     t = NothingTask(
         item_collection,
+        workdir=tmp_path / "test-task-download-nosuch-asset",
+        save_workdir=True,
     )
-    item = t.download_item_assets(t.items[0], assets=["nosuch_asset"]).to_dict()
-    # new item same as old item
-    assert item["assets"] == t.items[0].to_dict()["assets"]
+    item = t.download_item_assets(
+        t.items[0], config=DownloadConfig(include=["nosuch_asset"])
+    )
+
+    # new item has same assets hrefs as old item
+    assert [x.href for x in item.assets.values()] == [
+        x.href for x in t.items[0].assets.values()
+    ]
+
+
+def test_download_asset_dont_keep_existing(
+    tmp_path: Path, item_collection: dict[str, Any]
+) -> None:
+    t = NothingTask(
+        item_collection,
+        workdir=tmp_path / "test-task-download-nosuch-asset",
+        save_workdir=True,
+    )
+    item = t.download_item_assets(
+        t.items[0],
+        config=DownloadConfig(include=["nosuch_asset"]),
+        keep_non_downloaded=False,
+    )
+
+    # new item has no assets
+    assert item.assets == {}
 
 
 # @vcr.use_cassette(str(cassettepath / 'download_assets'))
 def test_download_item_asset(tmp_path: Path, item_collection: dict[str, Any]) -> None:
     t = NothingTask(item_collection, workdir=tmp_path / "test-task-download-item-asset")
-    item = t.download_item_assets(t.items[0], assets=["tileinfo_metadata"]).to_dict()
-    fname = item["assets"]["tileinfo_metadata"]["href"]
-    filename = Path(fname)
-    assert filename.is_file() is True
+    item = t.download_item_assets(
+        t.items[0], config=DownloadConfig(include=["tileinfo_metadata"])
+    )
+    assert Path(item.assets["tileinfo_metadata"].get_absolute_href()).is_file()
 
 
 def test_download_keep_original_filenames(
@@ -43,7 +72,11 @@ def test_download_keep_original_filenames(
         workdir=tmp_path / "test-task-download-item-asset",
     )
     item = t.download_item_assets(
-        t.items[0], assets=["tileinfo_metadata"], keep_original_filenames=True
+        t.items[0],
+        config=DownloadConfig(
+            include=["tileinfo_metadata"],
+            file_name_strategy=stac_asset.FileNameStrategy.FILE_NAME,
+        ),
     ).to_dict()
     fname = item["assets"]["tileinfo_metadata"]["href"]
     filename = Path(fname)
@@ -54,19 +87,25 @@ def test_download_item_asset_local(
     tmp_path: Path, item_collection: dict[str, Any]
 ) -> None:
     t = NothingTask(item_collection, workdir=tmp_path / "test-task-download-item-asset")
-    item = t.download_item_assets(t.items[0], assets=["tileinfo_metadata"])
-    fname = item.assets["tileinfo_metadata"].href
-    filename = Path(fname)
-    assert filename.is_file() is True
+    item = t.download_item_assets(
+        t.items[0], config=DownloadConfig(include=["tileinfo_metadata"])
+    )
+
+    assert (
+        Path(os.path.dirname(item.self_href)) / item.assets["tileinfo_metadata"].href
+    ).is_file()
+
     # Downloaded to local, as in prev test.
     # With the asset hrefs updated by the prev download, we "download" again to subdir
     item = t.download_item_assets(
-        item, assets=["tileinfo_metadata"], path_template="again/${collection}/${id}"
+        item=item,
+        config=DownloadConfig(include=["tileinfo_metadata"]),
+        path_template="again/${collection}/${id}",
     )
-    href = item.assets["tileinfo_metadata"].href
+    assert "again" in item.self_href
+    href = item.assets["tileinfo_metadata"].get_absolute_href()
     assert "again" in href
-    filename = Path(href)
-    assert filename.is_file() is True
+    assert Path(href).is_file()
 
 
 # @vcr.use_cassette(str(cassettepath / 'download_assets'))
@@ -77,12 +116,12 @@ def test_download_item_assets(tmp_path: Path, item_collection: dict[str, Any]) -
         save_workdir=True,
     )
     item = t.download_item_assets(
-        t.items[0], assets=["tileinfo_metadata", "granule_metadata"]
-    ).to_dict()
-    filename = Path(item["assets"]["tileinfo_metadata"]["href"])
-    assert filename.is_file() is True
-    filename = Path(item["assets"]["granule_metadata"]["href"])
-    assert filename.is_file() is True
+        t.items[0],
+        config=DownloadConfig(include=["tileinfo_metadata", "granule_metadata"]),
+    )
+
+    assert Path(item.assets["tileinfo_metadata"].get_absolute_href()).is_file()
+    assert Path(item.assets["granule_metadata"].get_absolute_href()).is_file()
 
 
 def test_download_items_assets(tmp_path: Path, item_collection: dict[str, Any]) -> None:
@@ -92,11 +131,11 @@ def test_download_items_assets(tmp_path: Path, item_collection: dict[str, Any]) 
         workdir=tmp_path / "test-task-download-items-assets",
         save_workdir=True,
     )
-    items = [i.to_dict() for i in t.download_items_assets(t.items, assets=[asset_key])]
-    filename = Path(items[0]["assets"][asset_key]["href"])
-    assert filename.is_file() is True
-    filename = Path(items[1]["assets"][asset_key]["href"])
-    assert filename.is_file() is True
+    items = t.download_items_assets(t.items, config=DownloadConfig(include=[asset_key]))
+
+    assert len(items) == 2
+    for item in items:
+        assert Path(item.assets[asset_key].get_absolute_href()).is_file()
 
 
 # @vcr.use_cassette(str(cassettepath / 'download_assets'))
@@ -108,8 +147,7 @@ def test_download_large_asset(tmp_path: Path, item_collection: dict[str, Any]) -
         save_workdir=True,
     )
     item = t.download_item_assets(
-        t.items[0], assets=["red"], requester_pays=True
-    ).to_dict()
-    filename = Path(item["assets"]["red"]["href"])
-    assert filename.is_file() is True
-    del t
+        t.items[0], config=DownloadConfig(s3_requester_pays=True, include=["red"])
+    )
+
+    assert Path(item.assets["red"].get_absolute_href()).is_file()